Refactor product API to use new response structures and implement soft delete functionality
- Updated API response structures to use `rest.SuccessResponse` and `rest.FailResponse` instead of `api.SuccessResponse` and `api.FailResponse`.
- Removed old API response definitions from Swagger and documentation files.
- Introduced new domain models for `Product` with timestamp and soft delete fields.
- Implemented new repository methods for product management with soft delete logic.
- Added SQLite migration for products table with new fields for timestamps and soft delete.
- Created new validation and response handling utilities for REST API.
- Updated product handler methods to reflect changes in response structures and validation.
- Modified request and response examples in HTTP request file for consistency with new API design.

+++ internal/product/request.go
@@ -3,6 +3,6 @@ package product
 
 type upsertRequest struct {
 	Name  string `json:"name" validate:"required" example:"Kopi Susu Gula Aren"`
-	Price *int    `json:"price" validate:"required,min=0" example:"15000"`
+	Price *int64    `json:"price" validate:"required,min=0" example:"15000"`
 	Stock *int    `json:"stock" validate:"required,min=0" example:"10"`
 }
