+++ internal/product/response.go
@@ -4,6 +4,6 @@ package product
 type productResponse struct {
 	ID    string `json:"id" example:"550e8400-e29b-41d4-a716-446655440000"`
 	Name  string `json:"name" example:"Kopi Susu Gula Aren"`
-	Price int    `json:"price" example:"15000"`
+	Price int64  `json:"price" example:"15000"`
 	Stock int    `json:"stock" example:"10"`
 }
